requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pandas>=2.0.0
openpyxl>=3.0.0
openai>=1.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0
pytest>=7.0.0
httpx>=0.24.0
pytest-cov>=4.0.0
ruff>=0.1.0
mypy>=1.0.0
black>=23.0.0
tenacity>=8.0.0